+++ backend/src/main/java/com/shopvr360/backend/repository/OdersRepository.java
@@ -1,11 +1,13 @@
 package com.shopvr360.backend.repository;
 
+import com.shopvr360.backend.entity.TableChitietdonhang;
 import com.shopvr360.backend.entity.TableDonhang;
 import com.shopvr360.backend.entity.TableTinh;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface OdersRepository extends JpaRepository<TableDonhang,Integer> {
+public interface OdersRepository extends JpaRepository<TableDonhang,Long > {
     List<TableDonhang> findAllByIdUser(Integer idUser);
+
 }
